requirements.txt
django==5.2.7
gunicorn==23.0.0
requests==2.31
sentry-sdk[django]==2.38.0
simplejson==3.20.1
django-cors-headers==4.9.0
django-crontab==0.7.1
django-health-check==3.20.0
channels==4.3.1
daphne==4.2.1
uvicorn==0.37.0
websockets==15.0.1
channels_redis==4.3.0
#pykerberos==1.2.4
pytz==2025.2
pycountry==24.6.1

meilisearch==0.37.0

pymongo[snappy]==4.15.1
# for using mongodb+srv:// URIs
dnspython==2.8.0
mongoengine==0.29.1

Babel==2.17.0

celery==5.5.3

# For tests
responses==0.25.8
urllib3==2.5.0
mongomock-5==5.0.1
behave-django==1.7.0
requests_mock==1.12.1
testcontainers==4.13.0

# For dev
pylint-django==2.6.1
ruff==0.13.1
line-profiler==5.0.0